adjust DIO, DAO period when mobility detected

+++ openwsn-fw/openstack/03b-IPv6/icmpv6rpl.h
@@ -172,7 +172,7 @@ void     icmpv6rpl_writeDODAGid(uint8_t* dodagid);
 uint8_t  icmpv6rpl_getRPLIntanceID(void);
 void     icmpv6rpl_getRPLDODAGid(uint8_t* address_128b);
 void     icmpv6rpl_setDIOPeriod(uint16_t dioPeriod);
-void     icmpv6rpl_setDAOPeriod(uint16_t daoPeriod);
+void     icmpv6rpl_setDAOPeriod(uint32_t daoPeriod);
 bool     icmpv6rpl_getPreferredParentIndex(uint8_t* indexptr);           // new DB
 bool     icmpv6rpl_getPreferredParentEui64(open_addr_t* addressToWrite); // new DB
 bool     icmpv6rpl_isPreferredParent(open_addr_t* address);              // new DB
